Get log directly from Webpack (#143)
The plugin sends log message through the socket
If the dashboard is used as it is now, this is simply ignored
It is however printed if the user doesn't provide a command

+++ bin/webpack-dashboard.js
@@ -17,22 +17,25 @@ program.option("-p, --port [port]", "Socket listener port");
 program.usage("[options] -- [script] [arguments]");
 program.parse(process.argv);
 
+var logFromChild = true;
+
 if (!program.args.length) {
-  program.outputHelp();
-  return;
+  logFromChild = false;
 }
 
-var command = program.args[0];
-var args = program.args.slice(1);
-var env = process.env;
-
-env.FORCE_COLOR = true;
-
-var child = spawn(command, args, {
-  env: env,
-  stdio: [null, null, null, null],
-  detached: true
-});
+if(logFromChild) {
+  var command = program.args[0];
+  var args = program.args.slice(1);
+  var env = process.env;
+  
+  env.FORCE_COLOR = true;
+  
+  var child = spawn(command, args, {
+    env: env,
+    stdio: [null, null, null, null],
+    detached: true
+  });
+}
 
 var dashboard = new Dashboard({
   color: program.color || "green",
@@ -42,30 +45,44 @@ var dashboard = new Dashboard({
 
 var port = program.port || 9838;
 var server = SocketIO(port);
-server.on("connection", function(socket) {
-  socket.on("message", function(message) {
-    dashboard.setData(message);
-  });
-});
 
 server.on("error", function(err) {
   console.log(err);
 });
 
-child.stdout.on("data", function (data) {
-  dashboard.setData([{
-    type: "log",
-    value: data.toString("utf8")
-  }]);
-});
 
-child.stderr.on("data", function (data) {
-  dashboard.setData([{
-    type: "log",
-    value: data.toString("utf8")
-  }]);
-});
 
-process.on("exit", function () {
-  process.kill(process.platform === "win32" ? child.pid : -child.pid);
-});
+if(logFromChild) {
+  server.on("connection", function(socket) {
+    socket.on("message", function(message) {
+      if(message.type !== "log") {
+        dashboard.setData(message);
+      }
+    });
+  });
+
+  child.stdout.on("data", function (data) {
+    dashboard.setData([{
+      type: "log",
+      value: data.toString("utf8")
+    }]);
+  });
+  
+  child.stderr.on("data", function (data) {
+    dashboard.setData([{
+      type: "log",
+      value: data.toString("utf8")
+    }]);
+  });
+
+  process.on("exit", function () {
+    process.kill(process.platform === "win32" ? child.pid : -child.pid);
+  });
+} else {
+  server.on("connection", function(socket) {
+    socket.on("message", function(message) {
+      dashboard.setData(message);
+    });
+  });
+}
+
